refactor TestGetSetValidatorSigningInfo test
